[FIX] l10n_mx_edi_landing: Fixed cfdi customs number traceback
Previous this commit you are able to create and set a landed cost through
the lot form view without validating the landed cost. This caused a traceback
later on when generating the CFDI or invoice document when trying to get the
formatted dates. Since a non validated landed cost might not have a date,
this raised an error.

This commit targets to fix this issue by expecting that a landed cost might
not have a date and also not allow through the form lot view to create a landed
cost

closes odoo/enterprise#95476

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>