[FIX] (website_)sale_renting: add to cart and night period
When adding a product from the shop page, the default renting dates,
based on the first rental price, are not computed, resulting on a
default period of 24 hours.
Getting back the same logic as adding to cart from the product page.

Night period cannot be mixed with other rental periods.

task-5065762

closes odoo/enterprise#94283

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>