[FIX] l10n_tr_reports: fix blank CSV export in General Ledger
Before this commit: After the refactor of the General Ledger in the
referenced commit, `l10n_tr_reports` no longer able to fetch the `aml_ids`.
This caused the CSV export of the General Ledger to be generated as blank.

Ref commit: https://github.com/odoo/enterprise/commit/235a5160d13296328b79e4092a8b88a733628268

After this commit: Ensured that `aml_ids` are properly retrieved, so that the
General Ledger CSV export contains the expected data.

closes odoo/enterprise#96117

X-original-commit: 09e0acebc1c96f471b97ae120ecbe86ace1db339
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>