[FIX] account_reports: GL additional groupby not foldable
Before when we added different groupby on the gl line than the default ones, that would not be unfoldable by default.

This is due to a default dict that set has_sublines to False by default and it would never be set to true for groupby other than the default ones.

closes odoo/enterprise#95824

X-original-commit: 6e591f5206309bf28e46e77de5df627c3483fff1
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>