[FIX] account_reports: Annotation tours are broken
yet again ...

It was possible for multiples loadAnnotations to be
present at the same time, losing the annotations from
all except one of them since we were reassigning the
annotations stored when fetching them.

o_embedded_actions wasn't acting perfectly, and we will
probably change it in the future so we are just going
back to the main working files kanban view and opening
the balances from there.

Linked Runbot Errors:
- https://runbot.odoo.com/odoo/runbot.build.error/231762
- https://runbot.odoo.com/odoo/runbot.build.error/232575
- https://runbot.odoo.com/odoo/runbot.build.error/232554

closes odoo/enterprise#95546

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>