[FIX] knowledge: only add one feff inside comment beacons
Context:
`\ufeff` characters are used in the editor to handle how the cursor behaves when
using arrow keys. In the case of knowledge comments, we want to have 2 cursor
positions at each comment boundary. I.e.: at the first beacon, we want a cursor
position before the anchor (outside of the comment), and one after (inside of
the comment).

Issue:
Since this [commit], empty links receive 2 `zwnbsp` nodes to allow the user to
type inside, instead of 1 previously.

However knowledge comments beacons are not editable and only need one `zwnbsp`,
since we never want to type inside. Furthermore, having 2 and nothing in between
them would cause an issue when their content is normalized, which may lead to a
traceback:

- `link_plugin` executes `selection.anchorNode.parentNode.normalize();` which
  merges 2 FEFF together. However the function `isZwnbsp` does return true for
  one text node with 2 FEFF characters. This means that during the next "add
  feffs" phase, new text nodes with one `FEFF` characters will be added,
  creating a growing loop generating new `FEFF` at every normalization phase.
- Furthermore, merging the nodes in one text node may not result in a new step
  in some cases, which prevent the `history_plugin` from registering the new
  node in time in its `nodeMap`, which may lead to a serialization traceback.

Resolution:
This commit stops using the deprecated `padLinkWithZwnbsp` method to ensure that
there are 3 `zwnbsp` characters for each beacon: 2 around it, and one inside.

[commit]: odoo/odoo@f0eaeb6

task-5046068

closes odoo/enterprise#95671

X-original-commit: 383f4ac54c28cc00d8cad54098ef62c588b6b47f
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>