[FIX] account_accountant: fix bank rec style problems
When you open a statement line from a reconciled move, it opens the bank
reconciliation widget with only the selected statement line, which is unfolded by default.

However, there are a few issues with this behavior, which are fixed in this commit:

1 - When entering the bank reconciliation widget, the initial line is unfolded.
If you remove the filter, all the other lines become unfolded as well.
This should not be the case; only the original line should remain unfolded.

2 - By default, the statement summary line is hidden. When the filter is removed,
the summary remains hidden. We now ensure the summary is displayed again when the filter is cleared.

3 - The Statement button on the statement line (which is meant to create a new statement)
doesn't make any sense when there is only one line. It is now hidden in this case.

task-5108118

X-original-commit: 275311feca8ab89fa083308d4d21442a2146f8e0
Part-of: odoo/enterprise#96670
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Igor Bertrand (igbe) <igbe@odoo.com>