[FIX] website_sale_renting_planning: take into account time off for unavailabilties
Before this commit, when an human resource is linked to rental product
thanks to his role and take a time off at a certain day. The user inside
ECommerce should be able to add that product when the resource is off
(since there is no resource available) at the period chosen. However,
that check does not take into account the time off taken by the
resources linked to that service rental product.

This commit makes sure the time off of the resources linked to the
product are correctly taken into account to know if the service rental
product is available at a specific period.

task-5075698

closes odoo/enterprise#102070

X-original-commit: 0feb032c5b005db903c2b05db3635476ddd1b177
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>