[PERF] quality_control: Avoid OOM for many QC
The fields `quality_check_todo` and `quality_check_fail` are calculated for
stock.picking records by iterating all checks of a stock.picking. This causes
a problem when clicking the "To Receive" button for a warehouse in the
inventory app, in case there are many transfers each with many quality checks.
The function will default to loading all data associated with quality checks in
memory through field prefetching. However, since quality checks have too much
data (particularly because of the HTML fields) associated with them, the cache
can quickly bloat causing an OOM error and crashing the worker.

This PR disables the prefetcher for quality checks before iterating them,
preventing this issue from happening since we only need very light fields in the
loop.

For a specific customer (opw-5025162), this was the case.

Benchmarks:

| No. stock.picking | avg no. quality checks | peak memory before | peak memory after |
| ----------------- | ---------------------- | ------------------ | ----------------- |
| 25                | 20                     | 2771 mb            | 235 mb            |

opw-5025162

closes odoo/enterprise#98304

X-original-commit: 00588d90f127648ad7a61a18451eaecbac5c71a7
Signed-off-by: Steve Van Essche <svs@odoo.com>