[FIX] knowledge: prevent knowledge comments in pre
There is a crash when inserting a knowledge comment in a /code block:

In this previous [task], insertion in `pre` elements was filtered to prevent
non-phrasing content from being inserted (as it is invalid per the html
specification). To prevent a crash, knowledge comments will be disabled in
`<pre>` elements, as they rely on `anchor` elements for the comment position in
the article body.

[task]: https://github.com/odoo/odoo/commit/216e9ebf02e8350b82ec676586d8bcaa2a446615

task-4984152

X-original-commit: 4561e449e8d5a455d50bfcc7364e4f6563655bbd
Part-of: odoo/enterprise#96047
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>