[FIX] {stock_barcode_,}quality_control: prevent quality check trigger on unpicked item
Issue:

In the barcode app, quality checks triggered at validation includes
quality checks related to unpicked products.

Steps to reproduce:

- Create a storable products product A tracked by SN
- Create a control points of type pass/fail on receipts control by
  quantity on product A
- Create and confirm a receipt with a move 2 x product A
- Open the receipt in the barcode app
- Scan product A > Scan SN001
- Click on Quality Check
> Both QC's are displayed to be processed

Expected behavior:

Only the QC related to the scanned SN should be processed as it is the
only unit that will be moved at validation.

Cause of the issue:

Only picked move lines are considered to be processed in the barcode app
However, the `check_quality` triggered by clicking on the quality check
button only check if the move related to the move line is picked:
https://github.com/odoo/enterprise/blob/9fe45b673c02a98e6dd6b3997f19a2018a76df09/quality_control/models/stock_picking.py#L64-L72

Fix:

Relying the `barcode_trigger` context key will ensure a uniform behavior
between the QC's displayed to be processed directly from the QC button
and from these displayed at validation since this context key is already
used at validation:
https://github.com/odoo/enterprise/blob/9fe45b673c02a98e6dd6b3997f19a2018a76df09/stock_barcode/static/src/models/barcode_model.js#L581-L590

Note we all changed the default return value of the `check_quality` from
`False` to `True` here:
https://github.com/odoo/enterprise/blob/9fe45b673c02a98e6dd6b3997f19a2018a76df09/quality_control/models/stock_picking.py#L71-L73
because this method is called in the `pre_action_done_hook` during the
`button_validate` of the picking:
https://github.com/odoo/odoo/blob/a97d3c772001f4f0b9df66d28c1c8f19358898e0/addons/stock/models/stock_picking.py#L1415-L1421
https://github.com/odoo/enterprise/blob/9fe45b673c02a98e6dd6b3997f19a2018a76df09/quality_control/models/stock_picking.py#L91-L96
and since a result that is not `True` is expected to be an action that
should be processed prior to validation, returning `False` would make
it impossible to proceed with a validation in case the `check_quality`
is called and there is no check to process.

opw-5010764

X-original-commit: 95d23c8b8e242091736fa21c0c3e271b2b3fc7e4
Part-of: odoo/enterprise#99799
Signed-off-by: William Henrotin (whe) <whe@odoo.com>