[FIX] esg: disable create action kanban/form of esg.database
Before this commit, when the user goes to Database menu inside ESG app,
the user can create a database if he wants. However, when he creates a
database, he might think it can add the link of the database to
automatically fetch the data, but this system is not implemented because
each database has its own formatting.

This commit disables the create action inside that menu to avoid
confusing the user. However, he can still create a database inside a
emission factor if he wants to classify his emission factors manually
created.

task-5117594

closes odoo/enterprise#95613

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>