[FIX] ai: add record details even if form view has no chatter
Purpose:
--------
Currently:
Creating a default prompt for "get help on a record" on a model that does
not inherit from `mail.trhead` does not work: it's the default "ask ai"
chat that is opened when clicking on the "ai" button in the systray from
this record's form view.
This means that the "prompt buttons" created are not shown, and that the
record info is not added in the prompt.
Also, clicking on this ai button in the systray from the res.users form
view results in a crash (one tries to fetch a thread because the model has a
"message_ids" field, but it's a related field and the model does not have
a thread)

With this commit:
- the info about the current record is always added to the context when
using the ai systray button from a form view, even if there's no chatter
- the "send as message" and "log note" are added from the launch chat
service, only if the model inherits from `mail.thread` (instead of relying
on the field `message_ids`)
- moved the call to open the ai chat back to the form controller to avoid
sending 2 events on the bus (currently systray sends an event to the form
controller which in turn sends an event with the model info to the systray),
and to make it easier to reuse this "open chat with agent" (shouldn't need
to always go through the systray for that)
- when creating default prompts for "ask ai for help", the models will
always be 'all models' (one does not want to have different "ask ai"s per
model - there's the 'get help on record' for form views that works per
model, but one does not really have actions to do on other view types)

Task-5109721

Part-of: odoo/enterprise#95638
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>