[FIX] documents: use an actual folder for Inbox KPI
After a discussion about what should be considered the inbox folder on
earlier versions, it appears that the kpi.provider:get_kpi_summary
method should return the count of documents in the folder having the
external identifier documents.documents_internal_folder, instead of
selecting the documents with the inbox tag.

As some documents might be referenced by foreign keys in demo data, we
also avoid to unlink the documents in the tests to avoid potential
errors in the test suites.

closes odoo/enterprise#97594

Runbot-build-error-id: 231787
Task-id: 5062431
X-original-commit: 6ed0b23af6b196302807a005b6bf559bcf101410
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Paul Morelle (pmo) <pmo@odoo.com>