[REF] iot: extract printer select into separate function
Before this commit, the printer selection wizard was tightly coupled to
the IoT report handler, and the call to send to the printer was
duplicated in the wizard and the handler.

After this commit, the printer selection wizard is contained in a
function that will always return the selected printers directly to the
caller, whether the dialog needs to be opened or not. The wizard is
simplified as a result and the print call always occurs in the handler.

This refactoring will allow other places to use the printer selection
wizard, namely for invoice printing.

X-original-commit: 241436732f8d0a8f92ca701c24eec22cb13e0c6b
Part-of: odoo/enterprise#95791
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>