[FIX] project_enterprise: fix row visibility when searching or filtering by user
Before this commit:
When searching for an assignee in the gantt view, the corresponding row was not
displayed if the user had no tasks assigned.
When searching for assignee even in any specific project it was showing all
unassigned task.

After this commit:
When searching by an assignee, the corresponding user row is now always visible
in the gantt view, even if the user has no assigned tasks.
Now when we do search for assignee under any specific project it will show
unassigned tasks of that project only.

task-5076701

closes odoo/enterprise#102072

X-original-commit: bfa3fa05d84d6e9ea1c82bd8033d99e37e786104
Related: odoo/odoo#239939
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>