[FIX] l10n_eu_oss_reports: consider country group of 1 country
To be able to deal with Spain with Canary Islands and mainland, we
have a country group that is Mainland Spain VAT, with Spain minus
several states (Canary basically).
People want to be able to use it for OSS.
But currently, if you have an entry with a tax with a fp with this
country group (and no country), you can't open your OSS Report.

So take the countries of the country group if there is none in the
fiscal position.

closes odoo/enterprise#95972

X-original-commit: 79a26b4fda5deb0637a69e13c6d5ed5addfb3914
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>