[FIX] {account,l10n_ro}_saft: include partners with 0 balance
We need to know for each partner if it is a customer or a supplier, even
more for Romania[^1] where it is enfored and validated.
> 1. If the element SD.P.22 CustomerID is reported with value ”0”
  (zero), then the element SD.P.23 SupplierID must be different from ”0”
  (zero), meaning the identity of the partner from which the purchase was
  made (conventionally considered ”supplier”) is reported. Else if SD.P.22
  CustomerID AND SD.P.23 SupplierID are concomitantly equal to ”0” (zero),
  then is return a semantic validation error. (CustomerID and SupplierID
  can not be concomitantly 0 (zero))

In order to fix this, we don't use the Partner Ledger anymore to query
the balance per partner because it is removing the partners with 0
balance automatically.
To keep it simple, we query manually and locally, allowing to reduce the
number of queries from 3 to 1 for that part.

For the performance, the `|=` operator done in a loop has also been
removed, keeping the time complexity in `O(n)` instead of `O(n²)`.

opw-5122910

[^1]: https://www.anaf.ro/anaf/internet/ANAF/despre_anaf/strategii_anaf/proiecte_digitalizare/saf_t

X-original-commit: 6b42f69181cc2607d16624f97074cfe4171dc80c
Part-of: odoo/enterprise#96355
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>