[FIX] l10n_mx_edi: Fix typo in fucntion.
In odoo, The standard function _check_move_constraints had a typo, as it was missing the last t (_check_move_constrains). When it was inherited in this module, the function was left with the typo. In version 19, this typo was fixed in the standard function, but the Mexican localization function was left as it was, rendering the function useless as it attempts to inherit a method that does not exist. In this PR, this typo is corrected so that the functionality also exists in 19.

Task-id: 5119381

I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr

closes odoo/enterprise#95825

Signed-off-by: Laurent Smet (las) <las@odoo.com>