[FIX] pos_enterprise: show custom attribute value on orderline
Steps to reproduce:
-------------------
1. Create a PoS product with an attribute of type "Radio", name it "X",
and add a value to it, name it "Y", with the "Free text" option
selected.
2. From PoS, click the product, for the attribute "X", select the value
   "Y", and enter some text in the text input area "blabla".
3. Order the product.

Observation: On the preparation display, only the attribute name and
value are display, but not the entered text, i.e. "X: Y" instead of "X:
Y: blabla".

Fix:
----
We now show the custom value (free text) if any. This matches what's
been shown on the receipt in PoS.

opw-5111768

closes odoo/enterprise#96280

X-original-commit: 3766caeffb4ac0fde18659c8e6c9c2e6e7b8a896
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Hadi El Yakhni (hael) <hael@odoo.com>