[FIX] esg: fetch ipcc database
Before this commit, sometimes the download button for IPCC database does
not work because the request made to IPCC database to fetch data, does
not return any data. The reason is because the table name used in
parameter changed after a certain time.

This commit gathers the table_name when we do extra rpc to fetch all
parameters to use to correctly export the data.

Part-of: odoo/enterprise#95879
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>