[FIX] account_reports: add lines without partner in initial balance
Issue:
The partner ledger does consider lines without partners when calculating the initial balance.

Steps to reproduce:
- Create an invoice in 2025
- Create an entry in 2025 without partner for the same amount
- Reconcile the two
- Open the partner ledger for 2025, everything is correct
- Change the dates to 2026, the amount of the initial balance ignores the entry but not the totals

Cause:
The method `_get_sums_without_partner` is called for the totals but not for the initial balance. Its purpose is to add the amounts of the lines without partners that were reconciled with lines with a partner.

Solution:
Call `_get_sums_without_partner()` in `_get_initial_balance_values()` add add the results before returning the initial balances.

As this is the same logic as `_query_partners()` we create a new method.

opw-5068790

closes odoo/enterprise#96341

X-original-commit: fd37a9a70a058140f07ec725e1377f59e56f3ed6
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>