[FIX] account_accountant: mark invoice partially paid in early payment discount
Problem: If partial payment is done for Invoice which is eligible for early
payment discount, Invoice is marked as paid and remaining amount is debited in
suspense account.

Cause: While preparing credit entry for receivable account, only eligibility of
Invoice for early payment discount was checked and whole residual amount of
Invoice  is credited instead of amount which is paid, and this is done for
partial payments also(which shouldn't be done).

After this commit: If partial payment is done for Invoice which is eligible for
early payment discount, Invoice is marked as partially paid and only amount
which is paid is credited from receivable account and nothing in suspense
account.

task-5128802

closes odoo/enterprise#97051

X-original-commit: bdb10ef3c531f1e1bd467761ddb3deb8e91ef9b6
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Pratham Nareshbhai Patel (ppat) <ppat@odoo.com>