[FIX] stock_barcode: stock quants not being updated
Issue:

In this bug, stock quants are not being updated when product_id
or location_id is updated.

To reproduce:
1- Create a db with demo database and barcode installed
2- Enable storage locations
3- Open barcode -> operations -> Internal transfers -> New
4- Add a product -> e.g. Drawer which there are quants in demo
5- As you see quants are not shown

Cause and Fix:

This is a partial backport of: #55917

`_compute_product_stock_quant_ids` should depend on
'product_id' and 'parent_location_id' to be recomputed when
product or source location is updated.

opw-5065624

closes odoo/enterprise#99016

X-original-commit: 30d2822b8219ac6d3f09982e16c40cae881bf515
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Mohammadmahdi Alijani (malj) <malj@odoo.com>