[FIX] account_iso20022: Missing field for DK banks
The denmark banks were refusing ISO20022 documents because there was a missing:
- "//PmtInf/PmtTpInf/LclInstrm/Cd"
or
- "//PmtInf/CdtTrfTxInf/PmtTpInf/LclInstrm/Cd"

One of those field should be filled by either 'ONCL' or 'SDCL' which means 'Over Night Clearing' or 'Same Day Clearing'.

To fix this we added a config parameter with a key: account_iso20022.local_instrument_code where we can set OCNL or SDCL to add the required field to the iso document. If nothing is set, the field will not be added.

opw-5073076

closes odoo/enterprise#96560

X-original-commit: d3c1af654c3bd6c334e1da7870a5a41a12f8e38b
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>