[FIX] account_accountant: lost match when adding account with tax
In the Bank reconciliation widget, users can click a button to set the
account to write off the remaining balance. However, if the chosen
account has a default tax set, the widget will lose any existing matches
with invoices.

Steps to reproduce:
- Have an account with a default tax
- Create a bill with a total
- Create a bank statement for a greater amount
- Open the bank reconciliation widget
- In the created statement, first add the bill, then click 'Set Account,' and
  choose the account with tax

Issue:
Bill matching will be lost.

This occurs because we remove and recreate the matching line, but we
don't keep the line to be reconciled.

opw-5002624

closes odoo/enterprise#96735

X-original-commit: 08f6c3806553d99c0403261c6e6aaf9d2a792f1d
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>