[FIX] l10n_mx_edi_*: Add missing readonly and invisible.
In refactor for l10n_mx_edi_stock, l10n_mx_edi_extra_weight was added to handle the weight of the driver and luggage. However, the condition to prevent the field from being edited once it has been stamped was not added, which could cause confusion if it could be edited later and a value different from that of the CFDI could be seen. The condition to make it readonly has been added, like the rest of the location fields in stock.picking.
In addition to this, when testing the flow, we also detected a conflict with Odoo Standard in the l10n_mx_edi_extended module. This module added the weight field again, now on the product accounting page. The standard weight field had a condition to make it invisible if the product had variants, but this one did not, which could cause confusion for customers, so these conditions were also added.

I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr

closes odoo/enterprise#95973

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>