[FIX] ai: prevent UnboundLocalError while messaging to ai
Currently an error is generated when a user tries to send a message to
AI while the current page is in `Physical Inventory`list view.

Steps:
- Install `Inventory`
- Go to Inventory > Operations > Physical Inventory
- Click the AI icon from the systray menu.
- Ask anything in AI >>> error generated

Error: `UnboundLocalError:cannot access local variable 'current_action'
 where it is not associated with a value`

This issue arises because in line [1] of the code, the variable `current_action`
is assigned a value inside an `if-elif` block based on `action.type`. However,
since the `Physical Inventory` page does not have any associated action, the
variable `current_action is` never set. Consequently, attempting to access
this variable results in an error.

This commit fixes the above issue by initializing the `current_action` variable
as `None` outside the `if-elif` block and adding handling for cases when
`current_action` is `None`.

[1] - https://github.com/odoo/enterprise/blob/2ea15cc9c7c5f114b3786b256c64e269b3e3a313/ai/models/ai_agent.py#L750-L755

sentry-6913801088

closes odoo/enterprise#96253

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>