[FIX] sign: delete role with radio button
Before this commit, when deleting a Signer role that already had a
radio button dropped in the Template, it would trigger a traceback
since the were trying to delete radio buttons individually. This
was problematic because radio buttons are deleted in chain, so when
deleting the first, we already delete all the remaining radio buttons
and its chain link.

After this commit, no issue is raised anymore when deleting a signer
role, as radio buttons already deleted are now skipped in the code.

task-5130925

closes odoo/enterprise#98425

X-original-commit: 9f774d6358828b468deb82634f04a3c6ef8b3e2e
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>