[FIX] ai: fix channels with prompt buttons but no default prompt
Purpose:
--------
When creating a chat that has "default prompts" but no instructions, the
agent always replies that the request can't be handled at the moment.

Issue arises because one adds `False` in the list of context, which leads
to a crash when the list of context is later converted into a string using
`join`.

Task-5109721

Part-of: odoo/enterprise#95638
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>