[FIX] account_sepa_direct_debit: mandate_pdf_file access
Removing the groups restriction from the mandate_pdf_file field in model sdd.mandate because it was causing issues when using the sdd.mandate.send wizard.
Any user who has access to the sdd.mandate model can use this wizard to print and send the record.
During this process, the system generates a PDF and stores it in the mandate_pdf_file binary field, linking the resulting attachment to the record.

The previous group restriction prevented users who were not part of the account.group_account_readonly group from sending the email with the attachment.
Even if the email was somehow sent, those users still couldn’t access the attachments they themselves had generated and sent.
With this change, any user who is allowed to send and print sdd.mandate records will also be able to generate and later access the corresponding attachments.

closes odoo/enterprise#96463

X-original-commit: 6cfd2672a9877b34478cedaf14538aba45ecad11
Related: odoo/odoo#230200
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>