[FIX] documents: use correct name in upload document invitation greeting
Currently, the email template inviting a user to upload a document
displays the document owner's name in the greeting instead of the requestee's.
This commit updates the template to display the correct name.

Steps to reproduce:
1. Log in as Mitchell Admin
2. Go in the Document app
3. Click on the New > Request
4. Enter a document name and select Joel Willis as requestee
5. Send the request by clicking on the "Request" button

=> The invitation email is correctly sent to Joel Willis but it says "Hello
Mitchell Admin" instead of "Hello Joel Willis"

Additionally, this commit ensures that the reminders and the invitation
emails are sent in the requestee's language rather than the sender's.

Task-5130591

closes odoo/enterprise#97092

X-original-commit: a71f9fc5f85550c731540a7baeaba0f5e3752e7a
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>