[PERF] helpdesk: improve updating partner fields
Issue:
When updating partner fields such as `email`, `name` and `phone` on the partner records,
the ticket fields are recomputed. During the recomputation, the `description` field is loaded into memory and with
many helpdesk tickets, this will cause a memory error.

Purpose of this PR:
To explicity fetch `partner_id` before computing partner fields on the helpdesk ticket.

Example:
With a partner record with 25,000 helpdesk tickets each with a description of varying text and images,
we hit the memory limit when trying to update the partner phone. With these changes, peak memory usage reached 271.4 MB.

Notes:
Originally wanted to avoid changing the field definition however this would require adding `with_context(prefetch_fields=False)`
to most of the compute methods on helpdesk ticket model.

opw-5069424

closes odoo/enterprise#96910

X-original-commit: f243de8d884311ed85bcd6959d208c56fc47f384
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Ian Dao (iada) <iada@odoo.com>