[FIX] account_reports: prevent error when updating template without recipients
Currently, an error occurs if the recipients of the statement report are removed
and the user tries to update the email template.

Steps to Reproduce:
1. Install Accounting.
2. Customers (list view) > select any record > Actions > Customer Statement > Send.
3. Remove recipients and try to update the email template.

Error:
KeyError - False

Cause:
The system tries to fetch mail field values from an empty recipient list,
leading to a KeyError.

Fix:
This commit recomputes the recipient partners before updating
the email subject and body from the template.

sentry-6918345730

closes odoo/enterprise#98340

X-original-commit: f4085d50a99fe880f5227d71944c53f67713b13d
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Aditi Patel (adip) <adip@odoo.com>