[FIX] website_sale_renting: move code from website_sale_renting_stock
Before this commit, when the user goes to the webshop to take a rental
service with rental service unavailable at a certain period, the system
does not block the user when the resource is not available during 2
hours in the period chosen by the user. The reason is because the hours
are not checked when website_sale_renting_stock is not installed.

This commit moves the code checking the time of the rental period made
in website_sale_renting_stock in website_sale_renting to be able to have
that verification for rental service used with planning to make sure the
system will prevent the user to add the product in his cart when the
resource is unavailable.

task-5123239

closes odoo/enterprise#102071

X-original-commit: 71e38b78d5f86e600a1ef054b264a04df94d8031
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>