[PERF] {l10n_*,account}_reports: Faster return generation
For more details with flamegraphs, see the
[PR](https://github.com/odoo/enterprise/pull/95760)

- batched _compute_company_ids on returns
- batched return to unlink in _generate_or_refresh_all_returns
- removed redondant call to _generate_or_refresh_all_returns in
  action_save_onboarding_fiscal_year since we want to call it only
  when a value is changed and that's already handled in the write on
  the company.
- batched _is_available_for on reports
- make only one write on the company (and one call to
  _generate_or_refresh_all_returns) when writing or create the fiscal
  year wizard.
- dont delete return created manually with a date before the date of
  the account_opening_date and not yet posted.

Detected from runbot error:
https://runbot.odoo.com/odoo/runbot.build.error/231463

A small workaround is needed on the company when writing the changes
from the fiscal year wizard. This is because related fields are writen
one by one, which then trigger multiple times
_generate_or_refresh_all_returns. To prevent this, we are batching
all the write on the company from the fiscal year wizard in one write.
The opening_date need to be handled separatly since it's not a related
field but still need to be writen on the company, if removed from the
vals during the create it's then not possible to save the wizard anymore.

Also fixed a bug where the returns manually created before the
account_opening_date would be deleted if they were not posted before
the next call to _try_create_returns_for_fiscal_year.

closes odoo/enterprise#95760

Related: odoo/odoo#231956
Signed-off-by: Julien Alardot (jual) <jual@odoo.com>