[FIX] l10n_mx_edi: Remove default invoice date override
At the moment, when an invoice that uses a CFDI is posted, the invoice
date (if not already existing) is set to the current date in the Mexico
City timezone.

This default behaviour is just weird, and even if there might have been
technical reasons for it in the past, these are no longer valid.

This can also cause the default invoice date to be different from the
default due date, which causes unexpected behaviour in tests.

We therefore remove this override.

runbot-233041

closes odoo/enterprise#96334

X-original-commit: ecdce137df0f1fcf6c5b067e3f80ff5807484b66
Signed-off-by: Laurent Smet (las) <las@odoo.com>