[FIX] documents: toggling debug mode through ctrl-k does not work
Issue:
In the documents app toggling debug mode through the command
palette (ctrl-k) does not work. Manually editing the URL works
as expected.

Cause:
A patch on `router.stateToUrl` for the documents app rebuilds the URL
to create clean, shareable links. When doing so, it attempted to
preserve the debug status by referencing the global `odoo.debug`.

This global object did not reflect client-side state changes. The
command palette correctly provides the intended future state to
the router. The patch was ignoring this incoming state and using the
outdated global value, causing the generated URL to be incorrect.

Solution:
Modify the patch to use the `state` argument provided to `stateToUrl`,
which contains the correct, up-to-date information for the navigation.

Task-5095056

closes odoo/enterprise#97721

X-original-commit: 2230a3eb0e6733cc8f21a969847df5e6fe89e409
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Jérôme Belpaire (beje) <beje@odoo.com>